# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HapiControl Internal Financial Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # USERS
...
  /caja:
    get:
      operationId: listCajaMovements
      tags: [caja]
      summary: List caja movements
      parameters:
        - name: executiveId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dateFrom
          in: query
          schema:
            type: ["string", "null"]
        - name: dateTo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Caja movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CajaMovement"
    post:
      operationId: createCajaMovement
      tags: [caja]
      summary: Create caja movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCajaMovementBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CajaMovement"
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - name: executiveId
          in: query
          schema:
            type: ["integer", "null"]
        - name: resolved
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Alerts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    patch:
      operationId: resolveAlert
      tags: [alerts]
      summary: Resolve alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"

  # DASHBOARD
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, executive, client, public]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, username, fullName, role, isActive, createdAt, updatedAt]
...
    CajaMovement:
      type: object
      properties:
        id:
          type: integer
        executiveId:
          type: integer
        executiveName:
          type: ["string", "null"]
        movementType:
          type: string
          enum: [collection, delivery, adjustment]
        amount:
          type: number
        description:
          type: ["string", "null"]
        relatedPaymentId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required: [id, executiveId, movementType, amount, createdAt]
    CreateCajaMovementBody:
      type: object
      properties:
        executiveId:
          type: integer
        movementType:
          type: string
          enum: [collection, delivery, adjustment]
        amount:
          type: number
        description:
          type: ["string", "null"]
        relatedPaymentId:
          type: ["integer", "null"]
      required: [executiveId, movementType, amount]
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        alertType:
          type: string
          enum: [due_today, overdue, broken_promise, renewal_opportunity, executive_inactive, unusual_movement, repeated_delay]
        severity:
          type: string
          enum: [low, medium, high]
        message:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: ["string", "null"]
        executiveId:
          type: ["integer", "null"]
        isResolved:
          type: boolean
        createdAt:
          type: string
      required: [id, alertType, severity, message, isResolved, createdAt]
...